fix(capi-ubuntu24): pin cloud-init deb package version
Newer version published on Ubuntu upstream stable channel introduced
differences in cloud-init behavior making AMI builds stale. Several
issues are around community/projects and there is not yet consensus on
how to move forward. At the same time, project is stale because of this
and a lot of users are affected.

Pinning cloud-init version to what we know for sure it's working to
unblock development. This is tested for both amd64/arm64 arch.

Signed-off-by: dntosas <ntosas@gmail.com>