chore: adding an option to bypass sysprep task that copies the grub file
This may be required by users who are using a custom source image and not using one straight form an ISO, for example building in OpenStack.
By providing a "false" flag, the process will skip over the sysprep task "Configure grub for non graphical consoles" which copies the grup file into place and causes any existing on to become redundant.